backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pymongo==4.6.0
dnspython==2.4.2
PyJWT==2.8.0
bcrypt==4.1.2
cloudinary==1.37.0
stripe==7.8.0
sendgrid==6.11.0
python-dateutil==2.8.2
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
